backend/requirements.txt
requests>=2.28
python-dotenv>=1.0
fastapi>=0.110
uvicorn[standard]>=0.24
pytest>=7.4
httpx>=0.26
